requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
click==8.1.7
fastapi==0.115.6
h11==0.14.0
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
pydantic==2.10.3
pydantic_core==2.27.1
python-multipart==0.0.19
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
uvicorn==0.32.1
